% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{CytoPipelineCheckApp}
\alias{CytoPipelineCheckApp}
\title{interactive visualization of flow cytometry data analysis pipeline
objects stored in cache}
\usage{
CytoPipelineCheckApp(dir = ".")
}
\arguments{
\item{dir}{the root directory into which the engine will look for existing
CytoPipeline experiments}
}
\value{
no return value
}
\description{
interactive visualization of flow cytometry data analysis pipeline
objects stored in cache
}
