% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipelineGUI-package.R
\docType{package}
\name{CytoPipelineGUI}
\alias{CytoPipelineGUI-package}
\alias{CytoPipelineGUI}
\title{CytoPipelineGUI package}
\description{
\code{CytoPipelineGUI} is the companion package of \code{CytoPipeline},
and is used for interactive visualization.
It implements two shiny applications :
\itemize{
\item a shiny app for interactive comparison of flow frames
that are the results of CytoProcessingSteps of the same
or different CytoPipeline experiments.
It is launched using the following statement: \code{CytoPipelineCheckApp()}
\item a shiny app for interactive visualization and manual adjustments of scale
transformation objects. It is launched using the following statement:
\code{ScaleTransformApp()}
}
}
\seealso{
\link[CytoPipeline:CytoPipeline-package]{CytoPipeline}

\link{CytoPipelineCheckApp}

\link{ScaleTransformApp}
}
\author{
\strong{Maintainer}: Philippe Hauchamps \email{philippe.hauchamps@uclouvain.be} (\href{https://orcid.org/0000-0003-2865-1852}{ORCID})

Authors:
\itemize{
  \item Laurent Gatto \email{laurent.gatto@uclouvain.be} (\href{https://orcid.org/0000-0002-1520-2268}{ORCID})
}

Other contributors:
\itemize{
  \item Dan Lin \email{dan.8.lin@gsk.com} [contributor]
}

}
\keyword{internal}
