% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ScaleTransformApp}
\alias{ScaleTransformApp}
\title{interactive display and modification of scale transform list}
\usage{
ScaleTransformApp(dir = ".")
}
\arguments{
\item{dir}{the root directory into which the engine will look for existing
CytoPipeline experiments}
}
\value{
nothing
}
\description{
this application allows the user to visualize a scale
transformation list, possibly amending it channel after channel, and save the
results on disk.
The needed input tranformation list and flow frame for visualization needs
to be read from a CytoPipeline experiments stored in cache.
}
